% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22.01.04_main1.4_filterKnownCovariates.R
\name{filterKnownCovariates}
\alias{filterKnownCovariates}
\title{Filter known covariates}
\usage{
filterKnownCovariates(
  knownCovariates,
  inferredCovariates,
  unadjustedR2_cutoff = 0.9,
  verbose = FALSE
)
}
\arguments{
\item{knownCovariates}{The known covariate matrix (must be observation by feature).}

\item{inferredCovariates}{The inferred covariate matrix, e.g., the top PCs (must be observation by feature as well). The rows (i.e., observations) in knownCovariates and inferredCovariates must match.}

\item{unadjustedR2_cutoff}{The cutoff value for unadjusted R squared.}

\item{verbose}{Logical. If \code{TRUE}, then some detailed messages are printed when the function is run.}
}
\value{
This function returns the filtered known covariates, i.e., the known covariates that should be kept.
}
\description{
This function filters out the known covariates that are captured well by the inferred covariates (unadjusted R squared greater than or equal to 0.9 by default). We use unadjusted R squared instead of adjusted R squared because we do not want to penalize for model complexity here.
}
\references{
Heather J. Zhou, Lei Li, Yumei Li, Wei Li, and Jingyi Jessica Li. PCA outperforms popular hidden variable inference methods for molecular QTL mapping. Genome Biology, 23(1):210, 2022.
}
