% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23.01.25_supp1.1_getResiduals.R
\name{getResiduals}
\alias{getResiduals}
\title{Get residuals}
\usage{
getResiduals(dataResponse, dataPredictors)
}
\arguments{
\item{dataResponse}{The matrix containing the response variables (must be observation by feature).}

\item{dataPredictors}{The matrix containing the predictors (must be observation by feature as well). The rows (i.e., observations) in dataResponse and dataPredictors must match.}
}
\value{
This function returns the residual matrix, which has the same dimensions as dataResponse.
}
\description{
Given dataResponse and dataPredictors, each a matrix, this function residualizes each variable in dataResponse against all the variables in dataPredictors using a multiple linear regression.
}
\references{
Heather J. Zhou, Lei Li, Yumei Li, Wei Li, and Jingyi Jessica Li. PCA outperforms popular hidden variable inference methods for molecular QTL mapping. Genome Biology, 23(1):210, 2022.
}
