% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22.01.09_func1.1_groupPredict.R
\name{groupPredict}
\alias{groupPredict}
\title{Group predict}
\usage{
groupPredict(
  dataResponse,
  dataPredictors,
  R2Type = c("unadjusted", "adjusted")
)
}
\arguments{
\item{dataResponse}{The matrix containing the response variables (must be observation by feature).}

\item{dataPredictors}{The matrix containing the predictors (must be observation by feature as well). The rows (i.e., observations) in dataResponse and dataPredictors must match.}

\item{R2Type}{Either "unadjusted" or "adjusted".}
}
\value{
This function returns a vector of unadjusted or adjusted squares, each corresponding to a response variable.
}
\description{
Given dataResponse and dataPredictors, each a matrix, this function predicts each variable in dataResponse with all the variables in dataPredictors using a multiple linear regression.
}
\references{
Heather J. Zhou, Lei Li, Yumei Li, Wei Li, and Jingyi Jessica Li. PCA outperforms popular hidden variable inference methods for molecular QTL mapping. Genome Biology, 23(1):210, 2022.
}
