% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22.01.04_main1.3_makeScreePlot.R
\name{makeScreePlot}
\alias{makeScreePlot}
\title{Make scree plot for PCA}
\usage{
makeScreePlot(
  prcompResult,
  labels,
  values,
  titleText = NULL,
  subtitleText = NULL,
  maxNumOfPCsToPlot = max(100, values),
  colors = NULL
)
}
\arguments{
\item{prcompResult}{The output from running \code{prcomp()}.}

\item{labels}{A string or vector of strings that will be used as the label(s) of the numbers of PCs in \code{values}.}

\item{values}{An integer or vector of integers representing the numbers of PCs chosen via different methods. \code{labels} and \code{values} must be nonempty, and the entries must match between the two vectors.}

\item{titleText}{Title text.}

\item{subtitleText}{Subtitle text.}

\item{maxNumOfPCsToPlot}{The maximum number of PCs to plot in the scree plot.}

\item{colors}{A string or vector of strings representing the colors to plot the numbers of PCs in \code{values} in.}
}
\value{
This function returns the produced scree plot.
}
\description{
This function makes a scree plot showing the numbers of PCs chosen via different methods.
}
\references{
Preprint.
}
