% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22.01.04_main1.2_runBE.R
\name{runBE}
\alias{runBE}
\title{The Buja and Eyuboglu (BE) algorithm for choosing K in PCA}
\usage{
runBE(
  X,
  B = 20,
  alpha = 0.05,
  mc.cores = min(B, parallel::detectCores() - 1),
  verbose = FALSE
)
}
\arguments{
\item{X}{The data matrix for running PCA (must be observation by feature).}

\item{B}{The number of permutations.}

\item{alpha}{The significance level.}

\item{mc.cores}{The number of cores used for parallel computing.}

\item{verbose}{Logical. If \code{TRUE}, then some detailed messages are printed when the function is run.}
}
\value{
This function returns a list consisting of the following items:
\item{pValues}{The p-values corresponding to the PCs.}
\item{alpha}{The significance level as inputted.}
\item{numOfPCsChosen}{The number of PCs chosen via BE.}
}
\description{
This function implements the BE algorithm, a permutation-based approach for choosing K in PCA. Intuitively, the BE algorithm retains PCs that explain more variance in the data than by random chance and discards those that do not.
}
\details{
The BE algorithm was originally proposed by Buja and Eyuboglu (1992). A detailed description of this method can be found in preprint. For reproducibility, make sure to change the RNG type (necessary unless \code{mc.cores} is 1) and set the seed before using this function (see README file for details).
}
\references{
Andreas Buja and Nermin Eyuboglu. 1992. “Remarks on Parallel Analysis.” Multivariate Behavioral Research 27 (4): 509–40.

Preprint.
}
