% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22.01.04_main1.1_runElbow.R
\name{runElbow}
\alias{runElbow}
\title{The elbow method for choosing K in PCA (based on distance to the diagonal line)}
\usage{
runElbow(X = NULL, prcompResult = NULL)
}
\arguments{
\item{X}{The data matrix for running PCA (must be observation by feature).}

\item{prcompResult}{The output from running \code{prcomp()} on \code{X}. Either \code{X} or \code{prcompResult} must be provided. If both are provided, then only \code{prcompResult} will be used. We recommend only providing \code{prcompResult} because it's faster.}
}
\value{
This function returns the number of PCs selected by maximizing the distance to the diagonal line.
}
\description{
This function implements an automatic elbow detection method (based on distance to the diagonal line).
}
\details{
A detailed description of this method can be found in preprint.
}
\references{
Preprint.
}
